feat: add possibility to specify size of operand in MED_ATTR*()
MED_ATTR* macros can take 4 or 5 arguments in this order:
   structname, structmember, (structmembersize,) attrname, medtype

The `structmembersize` allows specify the real size of
`structmember`, if its size differs from memory size of
implementing type. For example, kernel bitmaps are formed on x64
architecture by array of uin64 elements, but the real number of
used bits is up to programmer. This is the case of Medusa's virtual
spaces or subject's (object's) bitarray of events, that may be
triggered.

If `structmembersize` argument is not given, size of
`structmember`'s memory representation is used. This is the most
common case, so for simplicity, readability and backward
compatibility of l2 code remains also version of macros with four
arguments.