build: add Medusa's dependencies flag to Kconfig
Add explicit flag %SECURITY_MEDUSA_MET_DEPENDENCIES to Medusa's Kconfig
to check dependencies status during kernel configuration process. If the
dependencies are not met, an information how to resolve this situation
is shown.

For now, this is used to control the state, when any from the supported
hash function for Medusa's FUCK module is chosen as builtin.

As this flag is set only if %SECURITY_MEDUSA menuconfig entry is set (as
builtin) _and_ all dependencies are met, it can be used as dependency
control for all others Medusa's menu entries. Which is done with this
commit, too.