build: remove %SECURITY_MEDUSA_L4_CONSTABLE flag
Original version of Medusa is modular. That means, various parts of it
can be compiled as kernel modules. But, unfortunately, the LSM framework
does not allow a LSM loaded dynamically as a module.

Flag %SECURITY_MEDUSA_L4_CONSTABLE is the last modular residuum. Since
the L4 implementation is only one (for now) and stable, it should be
always present in running Medusa LSM.

For this reasons is the flag useless.